requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

fastapi
uvicorn[standard]
celery[redis]
redis
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-celery
opentelemetry-exporter-otlp

# Your ML deps
faster-whisper

# GPU-enabled PyTorch wheels
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
omegaconf==2.3.0
soundfile
numpy
huggingface_hub[hf_xet]
python-multipart
openai
